third_party/WebKit/LayoutTests/http/tests/security/mixedContent/data-url-script-in-data-iframe.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
async_test(t => {
    assert_equals(location.protocol, "https:");

    var f = document.createElement("iframe");
    window.addEventListener("message", t.step_func(e => {
        if (e.source !== f.contentWindow)
            return;
        assert_equals(e.data, "PASS");
        t.done();
    }));
    f.src = "data:text/html,<script src=\"data:application/javascript,parent.postMessage('PASS', '*');\"><\/script>";
    document.body.appendChild(f);
}, "'data:' URL frames can load 'data:' resources.");
</script>
</body>
